Fix a circular import
